+++ src/_scss/pages/search/filters/recipient/recipient.scss
@@ -115,59 +115,10 @@
         font-size: rem(30);
         text-align: center;
         margin-top: $global-margin;
+
         .recipient-filter-message-container__text {
-          font-size: 2rem;
-          padding-bottom: 2rem;
+            font-size: 2rem;
+            padding-bottom: 2rem;
         }
-    }
-
-    .recipient-filter-item-wrap {
-        padding: 0;
-
-        @include display(flex);
-        @include justify-content(center);
-        @include align-items(center);
-
-        .recipient-input {
-            @include flex(1 1 auto);
-            margin-right: rem(4);
-            font-size: $small-font-size;
-            color: $color-base;
-            line-height: rem(36);
-            height: rem(40);
-
-            @include placeholder {
-                color: $color-placeholder;
-            }
-        }
-
-        .recipient-submit {
-            @include flex(0 0 auto);
-            @include singleSubmit();
-        }
-    }
-
-    .shown {
-        min-height: rem(180);
-    }
-
-    .visually-hidden {
-        display: none;
-    }
-
-    .pop-typeahead {
-        .usa-da-typeahead {
-            margin-top: 0;
-        }
-
-        .funding {
-            .usa-da-typeahead {
-                margin: 0 0 rem(10);
-            }
-        }
-    }
-
-    .selected-filters {
-        @include selected-filter-wrap;
     }
 }
